+++ src/components/WebComponentProject/WebComponentProject.jsx
@@ -4,6 +4,7 @@ import { useMediaQuery } from "react-responsive";
 
 import Project from "../Editor/Project/Project";
 import MobileProject from "../Mobile/MobileProject/MobileProject";
+import Output from "../Editor/Output/Output";
 import { defaultMZCriteria } from "../../utils/DefaultMZCriteria";
 import Sk from "skulpt";
 import { setIsSplitView, setWebComponent } from "../../redux/EditorSlice";
@@ -19,7 +20,9 @@ const WebComponentProject = ({
   withProjectbar = false,
   withSidebar = false,
   sidebarOptions = [],
+  outputOnly = false,
 }) => {
+  const loading = useSelector((state) => state.editor.loading);
   const project = useSelector((state) => state.editor.project);
   const codeRunTriggered = useSelector(
     (state) => state.editor.codeRunTriggered,
@@ -37,6 +40,8 @@ const WebComponentProject = ({
     dispatch(setIsSplitView(false));
     dispatch(setWebComponent(true));
   }, [dispatch]);
+
+  console.log(`outputOnly: ${outputOnly}`);
 
   useEffect(() => {
     setCodeHasRun(false);
@@ -69,18 +74,26 @@ const WebComponentProject = ({
 
   return (
     <>
-      {isMobile ? (
-        <MobileProject
-          withSidebar={withSidebar}
-          sidebarOptions={sidebarOptions}
-        />
-      ) : (
-        <Project
-          forWebComponent={true}
-          withProjectbar={withProjectbar}
-          withSidebar={withSidebar}
-          sidebarOptions={sidebarOptions}
-        />
+      {!outputOnly &&
+        (isMobile ? (
+          <MobileProject
+            withSidebar={withSidebar}
+            sidebarOptions={sidebarOptions}
+          />
+        ) : (
+          <Project
+            forWebComponent={true}
+            withProjectbar={withProjectbar}
+            withSidebar={withSidebar}
+            sidebarOptions={sidebarOptions}
+          />
+        ))}
+      {outputOnly && (
+        <div className="embedded-viewer">
+          {loading === "success" && (
+            <Output embedded={true} browserPreview={false} />
+          )}
+        </div>
       )}
     </>
   );