+++ src/containers/WebComponentLoader.jsx
@@ -1,6 +1,13 @@
 import React, { useEffect, useState } from "react";
 import { useSelector, useDispatch } from "react-redux";
-import { disableTheming, setSenseHatAlwaysEnabled } from "../redux/EditorSlice";
+import {
+  disableTheming,
+  setEmbedded,
+  setSenseHatAlwaysEnabled,
+  stopCodeRun,
+  stopDraw,
+  triggerCodeRun,
+} from "../redux/EditorSlice";
 import WebComponentProject from "../components/WebComponentProject/WebComponentProject";
 import { useTranslation } from "react-i18next";
 import { setInstructions } from "../redux/InstructionsSlice";
@@ -22,6 +29,7 @@ import Style from "style-it";
 
 const WebComponentLoader = (props) => {
   const {
+    assetsIdentifier,
     authKey,
     identifier,
     code,
@@ -31,6 +39,7 @@ const WebComponentLoader = (props) => {
     withSidebar = false,
     sidebarOptions = [],
     theme,
+    outputOnly = false,
     embedded = false,
     hostStyles,
   } = props;
@@ -50,6 +59,7 @@ const WebComponentLoader = (props) => {
     (state) => state.editor.hasShownSavePrompt,
   );
   const saveTriggered = useSelector((state) => state.editor.saveTriggered);
+  const isEmbedded = useSelector((state) => state.editor.isEmbedded);
 
   const modals = useSelector((state) => state.editor.modals);
   const errorModalShowing = useSelector(
@@ -66,6 +76,23 @@ const WebComponentLoader = (props) => {
   const themeDefault = window.matchMedia("(prefers-color-scheme:dark)").matches
     ? "dark"
     : "light";
+
+  const [isRunCodeListenerAdded, setIsRunCodeListenerAdded] = useState(false);
+
+  const runCode = () => {
+    dispatch(stopCodeRun());
+    dispatch(stopDraw());
+    dispatch(triggerCodeRun());
+  };
+
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  useEffect(() => {
+    if (outputOnly && !isRunCodeListenerAdded) {
+      document.removeEventListener("outputOnly-runCode", runCode);
+      document.addEventListener("outputOnly-runCode", runCode);
+      setIsRunCodeListenerAdded(true);
+    }
+  });
 
   useEmbeddedMode(embedded);
 
@@ -92,8 +119,13 @@ const WebComponentLoader = (props) => {
     }
   }, [loading, project]);
 
+  if (embedded !== isEmbedded) {
+    dispatch(setEmbedded(embedded));
+  }
+
   useProject({
-    projectIdentifier: projectIdentifier,
+    assetsIdentifier,
+    projectIdentifier,
     code,
     accessToken: user?.access_token,
     loadRemix,
@@ -143,6 +175,7 @@ const WebComponentLoader = (props) => {
               withProjectbar={withProjectbar}
               withSidebar={withSidebar}
               sidebarOptions={sidebarOptions}
+              outputOnly={outputOnly}
             />
             {errorModalShowing && <ErrorModal />}
             {newFileModalShowing && <NewFileModal />}