+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+### Added
+
+- `assets_identifier` attribute for the web component (#901)
+- `code` attribute overriding content of `main.py`/`index.html` in the web component (#901)
+
 ### Changed
 
 - Made `p5` canvas responsive to the available space (#887)